# Conflicts:
#	src/api/Infrastructure/Shrooms.Infrastructure/CloudScheduler/AzureJob.cs
#	src/api/Infrastructure/Shrooms.Infrastructure/CloudScheduler/AzureJobScheduler.cs
#	src/api/Infrastructure/Shrooms.Infrastructure/CloudScheduler/Jobs/BirthdayNotifier.cs
#	src/api/Infrastructure/Shrooms.Infrastructure/CloudScheduler/Jobs/RecurringEvents.cs
#	src/api/Infrastructure/Shrooms.Infrastructure/ExcelGenerator/ExcelBuilder.cs
#	src/api/Infrastructure/Shrooms.Infrastructure/ExcelGenerator/IExcelBuilder.cs
#	src/api/Infrastructure/Shrooms.Infrastructure/Logger/Logger.cs
#	src/api/Infrastructure/Shrooms.Infrastructure/Shrooms.Infrastructure.csproj
#	src/api/Infrastructure/Shrooms.Infrastructure/Storage/AzureBlob/AzureStorage.cs
#	src/api/Infrastructure/Shrooms.Infrastructure/Storage/FileSystem/FileSystemStorage.cs
#	src/api/Infrastructure/Shrooms.IoC/Modules/BooksModule.cs
#	src/api/Infrastructure/Shrooms.IoC/Modules/ServicesModule.cs
#	src/api/Infrastructure/Shrooms.ModelMappings/Profiles/Kudos.cs
#	src/api/Infrastructure/Shrooms.ModelMappings/Profiles/Other.cs
#	src/api/Main/BuisnessLayer/Shrooms.DataTransferObjects/Models/GoogleBookApiService/ExternalBookInfo.cs
#	src/api/Main/BuisnessLayer/Shrooms.DataTransferObjects/Models/Kudos/KudosTypeDTO.cs
#	src/api/Main/BuisnessLayer/Shrooms.DataTransferObjects/Models/Kudos/NewKudosTypeDto.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Administration/AdministrationUsersService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Books/BookService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Email/Book/BooksNotificationService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Email/Book/IBooksNotificationService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Email/Kudos/KudosNotificationService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Email/Posting/CommentNotificationService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Email/Posting/PostNotificationService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Kudos/KudosService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Notifications/NotificationService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Picture/PictureService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Projects/ProjectsService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/UserService/UserService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Wall/Posts/PostService.cs
#	src/api/Main/BuisnessLayer/Shrooms.Domain/Services/Wall/WallService.cs
#	src/api/Main/BuisnessLayer/Shrooms.DomainServiceValidators/Validators/Books/BookMobileServiceValidator.cs
#	src/api/Main/BuisnessLayer/Shrooms.DomainServiceValidators/Validators/Books/BookServiceValidator.cs
#	src/api/Main/BuisnessLayer/Shrooms.DomainServiceValidators/Validators/Books/IBookMobileServiceValidator.cs
#	src/api/Main/BuisnessLayer/Shrooms.DomainServiceValidators/Validators/Books/IBookServiceValidator.cs
#	src/api/Main/DataLayer/Shrooms.DataLayer/DAL/ShroomsDbContext.cs
#	src/api/Main/DataLayer/Shrooms.EntityModels/Models/Kudos/KudosLog.cs
#	src/api/Main/DataLayer/Shrooms.EntityModels/Models/Notifications/NotificationsSettings.cs
#	src/api/Main/PresentationLayer/Shrooms.API/App_Start/Startup.Auth.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Controllers/BaseController.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Controllers/Book/BookController.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Controllers/CommentController.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Controllers/Kudos/KudosController.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Controllers/NotificationController.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Controllers/PictureController.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Controllers/PostController.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Controllers/Wall/WallController.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Filters/BasicAuthenticationAttribute.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Helpers/ClaimsHelper.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Hubs/NotificationHub.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Providers/ApplicationOAuthProvider.cs
#	src/api/Main/PresentationLayer/Shrooms.API/Web.config
#	src/api/Main/PresentationLayer/Shrooms.WebViewModels/Models/Users/Kudos/KudosLogsFilterViewModel.cs
#	src/api/Other/Shrooms.Constants/BusinessLayer/BusinessLayerConstants.cs
#	src/api/Other/Shrooms.Constants/Shrooms.Constants.csproj
#	src/api/Tests/Shrooms.Tests/Controllers/WebApi/BookControllerTests.cs
#	src/api/Tests/Shrooms.Tests/DomainService/AdministrationUsersServiceTests.cs
#	src/api/Tests/Shrooms.Tests/DomainService/BookServiceTests.cs
#	src/api/Tests/Shrooms.Tests/DomainService/KudosServiceTests.cs
#	src/api/Tests/Shrooms.Tests/DomainService/UserServiceTests.cs

+++ src/api/Infrastructure/Shrooms.Infrastructure/Storage/AzureBlob/AzureStorage.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Drawing;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.WindowsAzure.Storage;
@@ -27,27 +28,39 @@ namespace Shrooms.Infrastructure.Storage.AzureBlob
 
         public async Task RemovePicture(string blobKey, string tenantPicturesContainer)
         {
-            CloudBlockBlob blockBlob = GetBlockBlob(blobKey, tenantPicturesContainer);
+            var blockBlob = GetBlockBlob(blobKey, tenantPicturesContainer);
 
-            await blockBlob.DeleteAsync(DeleteSnapshotsOption.None, null, _blobRequestOptions, null);
+            if (blockBlob.Exists())
+            {
+                await blockBlob.DeleteAsync(DeleteSnapshotsOption.None, null, _blobRequestOptions, null);
+            }
+        }
+
+        public async Task UploadPicture(Image image, string blobKey, string mimeType, string tenantPicturesContainer)
+        {
+            var blockBlob = GetBlockBlob(blobKey, tenantPicturesContainer);
+            blockBlob.Properties.ContentType = mimeType;
+
+            using (var stream = new MemoryStream())
+            {
+                image.Save(stream, image.RawFormat);
+                stream.Position = 0;
+                await blockBlob.UploadFromStreamAsync(stream, null, _blobRequestOptions, null);
+            }
         }
 
         public async Task UploadPicture(Stream stream, string blobKey, string mimeType, string tenantPicturesContainer)
         {
-            CloudBlockBlob blockBlob = GetBlockBlob(blobKey, tenantPicturesContainer);
-
+            var blockBlob = GetBlockBlob(blobKey, tenantPicturesContainer);
             blockBlob.Properties.ContentType = mimeType;
-
             await blockBlob.UploadFromStreamAsync(stream, null, _blobRequestOptions, null);
         }
 
         private CloudBlockBlob GetBlockBlob(string blobKey, string containerName)
         {
-            CloudStorageAccount storageAccount = CloudStorageAccount.Parse(_settings.StorageConnectionString);
-
-            CloudBlobClient blobClient = storageAccount.CreateCloudBlobClient();
-
-            CloudBlobContainer container = blobClient.GetContainerReference(containerName);
+            var storageAccount = CloudStorageAccount.Parse(_settings.StorageConnectionString);
+            var blobClient = storageAccount.CreateCloudBlobClient();
+            var container = blobClient.GetContainerReference(containerName);
 
             return container.GetBlockBlobReference(blobKey);
         }
